add update provider route on agent
